divide test-final part, difference mostly in thread synchronization
